+++ internal/provider/provider.go
@@ -120,6 +120,7 @@ func (p *GigahostProvider) DataSources(_ context.Context) []func() datasource.Da
 	return []func() datasource.DataSource{
 		NewAccountDataSource,
 		NewSSHKeysDataSource,
+		NewDNSZonesDataSource,
 	}
 }